refactor(app-info): one action row for the info sheet and the details screen
Thor had two of these rows. Which actions an app offered depended on which
surface you reached it from:

  - Fix Store existed only in the info sheet, so the Apps tab's details
    screen could not re-point a sideloaded app at Play at all.
  - The Freezer membership toggle existed only in the details screen.
  - The same operations carried different labels and icons in each
    ("Launch"/"Open", "Kill"/"Force stop", "Cache"/"Clear cache",
    "Data"/"Clear data").
  - The sheet gated uninstalling Thor itself on a hardcoded
    "com.valhalla.thor" rather than BuildConfig.APPLICATION_ID.

Replace both with one AppActionRow widget. It decides whether an action is
applicable — privilege, system-app status, installer, feature settings — and
the caller decides what it does and whether a confirmation comes first. Two
callbacks are nullable because they are genuinely surface-specific: "Details"
is meaningless on the details screen, and the Freezer toggle needs membership
the caller may not know. A null callback hides its action; an action that can
never do anything useful is noise, not a hint.

The details screen gains Fix Store, and with it the risk dialog that has
always guarded it — checkAndProcessAction only routes Kill and ReinstallAll
through confirmation, so a Reinstall with no local dialog in front of it
would have reinstalled on a single tap.

The info sheet loses its header settings button; the row's Settings item is
now the single entry point instead of two a few dp apart.

Two behaviour notes:

  - Unfreezing from the info sheet now dismisses it, matching freeze. The
    sheet renders a snapshot AppInfo and cannot see the app come back, so
    leaving it open would show a stale "Unfreeze". Unobservable today —
    both hosts already clear their selection on every action.
  - The row is start-aligned rather than centred. It gains and loses items
    as privilege and installer state change, and centring slid every icon
    sideways when it did.

action_launch, action_data and cd_settings are now unreferenced; removed
across all five locales.
