+++ app/src/main/res/values/strings.xml
@@ -132,7 +132,6 @@
     <string name="status_suspended">SUSPENDED</string>
 
     <!-- App Actions -->
-    <string name="action_launch">Launch</string>
     <string name="action_share">Share</string>
     <string name="action_freeze">Freeze</string>
     <string name="action_unfreeze">Unfreeze</string>
@@ -140,7 +139,6 @@
     <string name="action_unsuspend">Unsuspend</string>
     <string name="action_kill">Kill</string>
     <string name="action_cache">Cache</string>
-    <string name="action_data">Data</string>
     <string name="action_uninstall">Uninstall</string>
     <string name="action_reinstall">Reinstall</string>
 
@@ -174,7 +172,6 @@
     <string name="cd_suspended">Suspended</string>
     <string name="cd_selected">Selected</string>
     <string name="cd_clear">Clear</string>
-    <string name="cd_settings">Settings</string>
     <string name="select_app_details">Select an app to view details</string>
 
     <!-- Permission Manager -->
